// apps/web/src/contracts/generated/sheet/Prim.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ClipPrim } from "./ClipPrim";
import type { GroupPrim } from "./GroupPrim";
import type { ImagePrim } from "./ImagePrim";
import type { ItemTag } from "./ItemTag";
import type { MapPrim } from "./MapPrim";
import type { PathPrim } from "./PathPrim";
import type { RectPrim } from "./RectPrim";
import type { TextPrim } from "./TextPrim";

export type Prim = { "type": "rect" } & RectPrim | { "type": "path" } & PathPrim | { "type": "text" } & TextPrim | { "type": "image" } & ImagePrim | { "type": "map" } & MapPrim | { "type": "pushClip" } & ClipPrim | { "type": "popClip" } & ItemTag | { "type": "pushGroup" } & GroupPrim | { "type": "popGroup" } & ItemTag;